Refactor middleware as pedastal-style interceptors
nREPL's use of `Transport` protocol as the extension mechanisism brings
incidental separation of logic between resquest/response handling. This
makes enter+exit pattern prefectly suited for middleware logic. It's a
shame that the middleware API wasn't designed like pedastal interceptors
in the first place.

This commit only refactors "enter" logic. That already brings the
benefit of makeing "eval" requests be printed immediately, rather than
having to wait for eval to complete and be printed along with the
response.